xxxxxx

+++ MCHexBOT/Protocol/Pakets/Client/Play/EntityRotationPaket.cs
@@ -5,23 +5,23 @@ namespace MCHexBOT.Pakets.Client.Play
     internal class EntityRotationPaket : IPaket
     {
         public int EntityId { get; set; }
-        public int Yaw { get; set; }
-        public int Pitch { get; set; }
+        public byte Yaw { get; set; }
+        public byte Pitch { get; set; }
         public bool OnGround { get; set; }
 
         public void Decode(MinecraftStream minecraftStream)
         {
             EntityId = minecraftStream.ReadVarInt();
-            Yaw = minecraftStream.ReadByte();
-            Pitch = minecraftStream.ReadByte();
+            Yaw = (byte)minecraftStream.ReadByte();
+            Pitch = (byte)minecraftStream.ReadByte();
             OnGround = minecraftStream.ReadBool();
         }
 
         public void Encode(MinecraftStream minecraftStream)
         {
             minecraftStream.WriteVarInt(EntityId);
-            minecraftStream.WriteByte((byte)Yaw);
-            minecraftStream.WriteByte((byte)Pitch);
+            minecraftStream.WriteByte(Yaw);
+            minecraftStream.WriteByte(Pitch);
             minecraftStream.WriteBool(OnGround);
         }
     }
